requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
black==26.3.1
blis==1.0.2
bs4==0.0.2
catalogue==2.0.10
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.1
cloudpathlib==0.23.0
colorama==0.4.6
confection==0.1.5
cymem==2.0.13
distro==1.9.0
dnspython==2.8.0
docx2txt==0.9
email-validator==2.3.0
fastapi==0.135.3
fastapi-cli==0.0.24
fulltext==0.5
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
isort==8.0.1
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
langcodes==3.5.1
language_data==1.4.0
lxml==6.0.2
marisa-trie==1.4.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
murmurhash==1.0.15
mypy_extensions==1.1.0
numpy==2.4.4
openai==2.30.0
orjson==3.11.8
packaging==26.0
passlib==1.7.4
pathspec==1.0.4
pillow==12.2.0
platformdirs==4.9.4
preshed==3.0.13
psycopg2-binary==2.9.11
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.20.0
PyJWT==2.12.1
PyPDF2==3.0.1
python-docx==1.2.0
python-dotenv==1.2.2
python-multipart==0.0.22
python-pptx==1.0.2
pytokens==0.4.1
PyYAML==6.0.3
regex==2026.3.32
requests==2.33.1
rich==14.3.3
rich-toolkit==0.19.7
setuptools==82.0.1
shellingham==1.5.4
six==1.17.0
smart_open==7.5.1
sniffio==1.3.1
soupsieve==2.8.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.48
srsly==2.5.3
starlette==1.0.0
structlog==25.5.0
thinc==9.1.1
tqdm==4.67.3
typer==0.24.1
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
ujson==5.12.0
urllib3==2.6.3
uvicorn==0.42.0
wasabi==1.1.3
watchfiles==1.1.1
weasel==0.4.3
websockets==16.0
wrapt==2.1.2
xlrd==2.0.2
xlsxwriter==3.2.9
